requirements.txt
torch>=2.0.0
git+https://github.com/openai/CLIP.git
datasets>=2.14.0
numpy>=1.24.0
Pillow>=9.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0
jupyter>=1.0.0
